Repurpose shared_uber_future as cuda::shared_future Eliminate the old version of cuda::shared_future
