+++ agency/cuda/detail/future/shared_future.hpp
@@ -17,7 +17,7 @@
 #pragma once
 
 #include <agency/detail/config.hpp>
-#include <agency/cuda/detail/future/async_future.hpp>
+#include <agency/cuda/detail/future/future.hpp>
 #include <type_traits>
 #include <memory>
 
@@ -27,30 +27,21 @@ namespace cuda
 {
 
 
+// XXX TODO port this class's members to __host__ __device__
+//          by introducing a CUDA-compatible implementation of shared_ptr
 template<class T>
 class shared_future
 {
   private:
-    std::shared_ptr<async_future<T>> underlying_future_;
-
-    detail::event& event()
-    {
-      return underlying_future_->event();
-    } // end event()
-
-    auto data() const ->
-      decltype(underlying_future_->data())
-    {
-      return underlying_future_->data();
-    }
+    std::shared_ptr<agency::cuda::future<T>> underlying_future_;
 
   public:
     shared_future() = default;
 
     shared_future(const shared_future&) = default;
 
-    shared_future(async_future<T>&& other)
-      : underlying_future_(std::make_shared<async_future<T>>(std::move(other)))
+    shared_future(agency::cuda::future<T>&& other)
+      : underlying_future_(std::make_shared<agency::cuda::future<T>>(std::move(other)))
     {}
 
     shared_future(shared_future&& other) = default;
@@ -76,36 +67,23 @@ class shared_future
       underlying_future_->wait();
     }
 
-  private:
-    void convert_data_to_result_of_get(agency::detail::unit_ptr)
-    {
-    }
-
-    template<class U>
-    const U& convert_data_to_result_of_get(U* ptr)
-    {
-      return *ptr;
-    }
-
-  public:
     auto get() ->
-      decltype(this->convert_data_to_result_of_get(this->data()))
+      decltype(underlying_future_->get_ref())
     {
-      wait();
-      return this->convert_data_to_result_of_get(this->data());
+      return underlying_future_->get_ref();
     } // end get()
 
     template<class... Args,
              class = typename std::enable_if<
-               detail::is_constructible_or_void<T,Args...>::value
+               agency::cuda::detail::is_constructible_or_void<T,Args...>::value
              >::type>
     static shared_future make_ready(Args&&... args)
     {
-      return async_future<T>::make_ready(std::forward<Args>(args)...).share();
+      return agency::cuda::future<T>::make_ready(std::forward<Args>(args)...).share();
     }
 
     template<class Function>
-    async_future<
+    agency::cuda::future<
       agency::detail::result_of_continuation_t<
         typename std::decay<Function>::type,
         shared_future
@@ -120,6 +98,13 @@ class shared_future
 
       return underlying_future_->then_and_leave_valid(f);
     }
+
+    template<class Function, class Factory, class Shape, class IndexFunction, class OuterFactory, class InnerFactory>
+    agency::cuda::future<typename std::result_of<Factory(Shape)>::type>
+      bulk_then(Function f, Factory result_factory, Shape shape, IndexFunction index_function, OuterFactory outer_factory, InnerFactory inner_factory, agency::cuda::gpu_id gpu)
+    {
+      return underlying_future_->bulk_then_and_leave_valid(f, result_factory, shape, index_function, outer_factory, inner_factory, gpu);
+    }
 };
 
 
@@ -127,6 +112,15 @@ class shared_future
 // requires the definition of one of shared_future<T>'s ctors
 template<class T>
 shared_future<T> async_future<T>::share()
+{
+  return shared_future<T>(std::move(*this));
+}
+
+
+// implement future<T>::share() here because this implementation
+// requires the definition of one of shared_future<T>'s ctors
+template<class T>
+shared_future<T> future<T>::share()
 {
   return shared_future<T>(std::move(*this));
 }
