+++ agency/cuda/future.hpp
@@ -4,4 +4,5 @@
 #include <agency/cuda/detail/future/async_future.hpp>
 #include <agency/cuda/detail/future/deferred_future.hpp>
 #include <agency/cuda/detail/future/future.hpp>
+#include <agency/cuda/detail/future/shared_future.hpp>
