+++ test_shared_future.cu
@@ -1,6 +1,6 @@
 #include <cassert>
-#include <agency/cuda/detail/future/shared_future.hpp>
 #include <iostream>
+#include <agency/cuda/future.hpp>
 
 int main()
 {
@@ -52,7 +52,7 @@ int main()
     assert(f2.valid());
     assert(f2.is_ready());
 
-    auto f3 = f2.then([] __device__ (int& arg)
+    auto f3 = f2.then([](int& arg)
     {
       return arg + 7;
     });
@@ -60,7 +60,7 @@ int main()
     assert(f2.valid()); // f2 is a shared_future and should still be valid after a .then()
     assert(f3.valid()); // f3 is a future and should be valid
 
-    auto f4 = f3.then([] __device__ (int& arg)
+    auto f4 = f3.then([](int& arg)
     {
       return arg + 42;
     });
