+++ testing/cuda/test_async_future_cast.cu
@@ -11,10 +11,10 @@ int main()
   {
     // construction of future<void> from empty types
 
-    using future_type1 = agency::cuda::future<empty>;
-    future_type1 f1 = agency::cuda::make_ready_future(empty());
+    using future_type1 = agency::cuda::async_future<empty>;
+    future_type1 f1 = agency::cuda::make_ready_async_future(empty());
 
-    using future_type2 = agency::cuda::future<void>;
+    using future_type2 = agency::cuda::async_future<void>;
     future_type2 f2(std::move(f1));
 
     assert(!f1.valid());
@@ -23,10 +23,10 @@ int main()
 
   {
     // empty -> void via future_traits
-    using future_type1 = agency::cuda::future<empty>;
-    future_type1 f1 = agency::cuda::make_ready_future(empty());
+    using future_type1 = agency::cuda::async_future<empty>;
+    future_type1 f1 = agency::cuda::make_ready_async_future(empty());
 
-    using future_type2 = agency::cuda::future<void>;
+    using future_type2 = agency::cuda::async_future<void>;
     future_type2 f2 = agency::future_traits<future_type1>::cast<void>(f1);
 
     assert(!f1.valid());
@@ -35,10 +35,10 @@ int main()
 
   {
     // empty -> void via executor_traits
-    using future_type1 = agency::cuda::future<empty>;
-    future_type1 f1 = agency::cuda::make_ready_future(empty());
+    using future_type1 = agency::cuda::async_future<empty>;
+    future_type1 f1 = agency::cuda::make_ready_async_future(empty());
 
-    using future_type2 = agency::cuda::future<void>;
+    using future_type2 = agency::cuda::async_future<void>;
 
     agency::cuda::grid_executor exec;
 
@@ -50,10 +50,10 @@ int main()
 
   {
     // unsigned int -> int via future_traits
-    using future_type1 = agency::cuda::future<unsigned int>;
-    future_type1 f1 = agency::cuda::make_ready_future(13u);
+    using future_type1 = agency::cuda::async_future<unsigned int>;
+    future_type1 f1 = agency::cuda::make_ready_async_future(13u);
 
-    using future_type2 = agency::cuda::future<int>;
+    using future_type2 = agency::cuda::async_future<int>;
     future_type2 f2 = agency::future_traits<future_type1>::cast<int>(f1);
 
     // XXX fut.then() needs to invalidate fut
@@ -63,10 +63,10 @@ int main()
 
   {
     // unsigned int -> int via executor_traits
-    using future_type1 = agency::cuda::future<unsigned int>;
-    future_type1 f1 = agency::cuda::make_ready_future(13u);
+    using future_type1 = agency::cuda::async_future<unsigned int>;
+    future_type1 f1 = agency::cuda::make_ready_async_future(13u);
 
-    using future_type2 = agency::cuda::future<int>;
+    using future_type2 = agency::cuda::async_future<int>;
 
     agency::cuda::grid_executor exec;
