+++ testing/cuda/test_grid_executor_member_detection.cu
@@ -58,7 +58,7 @@ struct int_factory
 int main()
 {
   using executor_type = agency::cuda::grid_executor;
-  using int_future_type = agency::cuda::future<int>;
+  using int_future_type = agency::cuda::async_future<int>;
 
   static_assert(
     agency::detail::new_executor_traits_detail::has_multi_agent_then_execute_with_shared_inits_returning_user_specified_container<executor_type, functor_returning_int, container_factory<container<int>>, int_future_type, int_factory, int_factory>::value,
