+++ testing/cuda/test_when_all_execute.cu
@@ -77,8 +77,8 @@ int main()
 
   {
     // int, float -> (int, float)
-    agency::cuda::future<int>   f1 = agency::cuda::make_ready_future<int>(7);
-    agency::cuda::future<float> f2 = agency::cuda::make_ready_future<float>(13);
+    auto f1 = agency::cuda::make_ready_async_future<int>(7);
+    auto f2 = agency::cuda::make_ready_async_future<float>(13);
 
     auto f3 = when_all_execute([] __device__ (agency::cuda::grid_executor::index_type idx, int& past_arg1, float& past_arg2, int& outer_arg, int& inner_arg)
     {
@@ -101,8 +101,8 @@ int main()
 
   {
     // int, void -> int
-    agency::cuda::future<int>  f1 = agency::cuda::make_ready_future<int>(7);
-    agency::cuda::future<void> f2 = agency::cuda::make_ready_future();
+    auto f1 = agency::cuda::make_ready_async_future<int>(7);
+    auto f2 = agency::cuda::make_ready_async_future();
 
     auto f3 = when_all_execute([] __device__ (agency::cuda::grid_executor::index_type idx, int& past_arg, int& outer_arg, int& inner_arg)
     {
@@ -124,8 +124,8 @@ int main()
 
   {
     // void, int -> int
-    agency::cuda::future<void> f1 = agency::cuda::make_ready_future();
-    agency::cuda::future<int>  f2 = agency::cuda::make_ready_future<int>(7);
+    auto f1 = agency::cuda::make_ready_async_future();
+    auto f2 = agency::cuda::make_ready_async_future<int>(7);
 
     auto f3 = when_all_execute([] __device__ (agency::cuda::grid_executor::index_type idx, int& past_arg, int& outer_arg, int& inner_arg)
     {
@@ -147,7 +147,7 @@ int main()
 
   {
     // void -> void
-    agency::cuda::future<void> f1 = agency::cuda::make_ready_future();
+    auto f1 = agency::cuda::make_ready_async_future();
 
     auto f3 = when_all_execute([] __device__ (agency::cuda::grid_executor::index_type idx, int& outer_arg, int& inner_arg)
     {
@@ -163,8 +163,8 @@ int main()
 
   {
     // void, void -> void
-    agency::cuda::future<void> f1 = agency::cuda::make_ready_future();
-    agency::cuda::future<void> f2 = agency::cuda::make_ready_future();
+    auto f1 = agency::cuda::make_ready_async_future();
+    auto f2 = agency::cuda::make_ready_async_future();
 
     auto f3 = when_all_execute([] __device__ (agency::cuda::grid_executor::index_type idx, int& outer_arg, int& inner_arg)
     {
